fix: move out /programs/getPrograms and fold programApi.ts into programsApi.ts
