+++ frontend/src/pages/CourseSelector/CourseBanner/CourseBanner.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useQuery } from '@tanstack/react-query';
 import { Typography } from 'antd';
 import { CoursesResponse } from 'types/userResponse';
-import { fetchAllDegrees } from 'utils/api/programApi';
+import { fetchAllDegrees } from 'utils/api/programsApi';
 import { getUserDegree } from 'utils/api/userApi';
 import CourseSearchBar from 'components/CourseSearchBar';
 import { useAppDispatch } from 'hooks';
