+++ frontend/src/utils/api/degreeApi.ts
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import { Programs } from 'types/api';
 import { DegreeWizardPayload } from 'types/degreeWizard';
 import { withAuthorization } from './auth';
 
@@ -12,9 +11,4 @@ export const setupDegreeWizard = async (token: string, wizard: DegreeWizardPaylo
     console.error('Error resetting degree at setupDegreeWizard: ', err);
     throw err;
   }
-};
-
-export const getAllDegrees = async (): Promise<Record<string, string>> => {
-  const res = await axios.get<Programs>('/programs/getPrograms');
-  return res.data.programs;
 };
